requirements.txt
netaddr==0.7.19
pyserial==3.4
scapy==2.4.3
